feat: custom view architecture
